refactor: 도메인 내부에 엔티티별 하나의 디렉토리로 구분 #16
