+++ src/main/java/KUSITMS/WITHUS/domain/application/position/entity/Position.java
@@ -1,4 +1,4 @@
-package KUSITMS.WITHUS.domain.application.entity;
+package KUSITMS.WITHUS.domain.application.position.entity;
 
 import KUSITMS.WITHUS.domain.organization.entity.Organization;
 import KUSITMS.WITHUS.global.common.BaseEntity;
