+++ src/main/java/KUSITMS/WITHUS/domain/application/position/repository/PositionRepositoryImpl.java
@@ -1,7 +1,7 @@
-package KUSITMS.WITHUS.domain.application.repository;
+package KUSITMS.WITHUS.domain.application.position.repository;
 
 
-import KUSITMS.WITHUS.domain.application.entity.Position;
+import KUSITMS.WITHUS.domain.application.position.entity.Position;
 import KUSITMS.WITHUS.global.exception.CustomException;
 import KUSITMS.WITHUS.global.exception.ErrorCode;
 import com.querydsl.jpa.impl.JPAQueryFactory;
