+++ src/main/java/KUSITMS/WITHUS/domain/application/position/service/PositionServiceImpl.java
@@ -1,9 +1,9 @@
-package KUSITMS.WITHUS.domain.application.service;
+package KUSITMS.WITHUS.domain.application.position.service;
 
-import KUSITMS.WITHUS.domain.application.dto.PositionRequestDTO;
-import KUSITMS.WITHUS.domain.application.dto.PositionResponseDTO;
-import KUSITMS.WITHUS.domain.application.entity.Position;
-import KUSITMS.WITHUS.domain.application.repository.PositionRepository;
+import KUSITMS.WITHUS.domain.application.position.dto.PositionRequestDTO;
+import KUSITMS.WITHUS.domain.application.position.dto.PositionResponseDTO;
+import KUSITMS.WITHUS.domain.application.position.entity.Position;
+import KUSITMS.WITHUS.domain.application.position.repository.PositionRepository;
 import KUSITMS.WITHUS.domain.organization.entity.Organization;
 import KUSITMS.WITHUS.domain.organization.repository.OrganizationRepository;
 import lombok.RequiredArgsConstructor;
